requirements.txt
pypdf>=4.2.0
python-docx>=1.1.0
